Fix TCP out of order
When TCP packets in the PCAP aren't recorded in the right order, the
payload would become corrupted, because data was directly appended the
end of the current payload and no 'filling' was inserted.
This data would then later be overwritten by another chunk which would
be written to the correct position for this chunk (but the packet that
was out of order falsely occupied this offset beforehand).

The fix consists of extending the payload with zeroes, up to the
position where the current chunk belongs. These zeroes would then later
be overwritten by the data that belongs in this place (if the packet
arrives).

@daniel-plohmann helped discover this bug.